Initial code to make watch also use iCloud key value store
