--- daysleft WatchKit Extension/WatchDaysLeftModel.swift
@@ -1,56 +1,0 @@
-//
-//  WatchDaysLeftModel.swift
-//  daysleft
-//
-//  Created by John Pollard on 03/10/2015.
-//  Copyright © 2015 Brave Location Software. All rights reserved.
-//
-
-import Foundation
-import WatchConnectivity
-import ClockKit
-
-class WatchDaysLeftModel: AppSettingsDataManager {
-    public init() {
-        super.init()
-        self.updateComplications()
-    }
-    
-    /// Send updated settings to watch
-    override func initialiseiCloudSettings() {
-        // Don't setup iCloud on watch!
-    }
-    
-    /// Write settings to iCloud store
-    override func writeSettingToiCloudStore(_ value: AnyObject, key: String) {
-        // Don't write settings to iCloud when on watch!
-    }
-    
-    /// Send initial settings to watch
-    override func pushAllSettingsToWatch() {
-        // Don't send settings to watch when on watch!
-    }
-        
-    override func session(_ session: WCSession, didReceiveUserInfo userInfo: [String: Any]) {
-        super.session(session, didReceiveUserInfo: userInfo)
-        self.updateComplications()
-    }
-
-    override func session(_ session: WCSession, didReceiveUpdate receivedApplicationContext: [String: AnyObject]) {
-        super.session(session, didReceiveUpdate: receivedApplicationContext)
-        self.updateComplications()
-    }
-    
-    func updateComplications() {
-        print("Updating complications...")
-        
-        let complicationServer = CLKComplicationServer.sharedInstance()
-        let activeComplications = complicationServer.activeComplications
-        
-        if (activeComplications != nil) {
-            for complication in activeComplications! {
-                complicationServer.reloadTimeline(for: complication)
-            }
-        }
-    }
-}
